Disambiguate license
I noticed that there isn't a clear license for this work, and want to make sure we respect your work and comply with any terms. I also may have some quick bugfixes in the not too distant future. 
I'm proposing a license, and a fairly permissive one at that - but primarily just to move the conversation forward. If 2 Clause BSD is acceptable - merge is pretty simple. If not, then picking another is fine.